Move the data/utility.[h|cc] to bridge/utility.[h|cc]
