+++ mobile/library/cc/key_value_store.cc
@@ -1,6 +1,6 @@
 #include "library/cc/key_value_store.h"
 
-#include "library/common/data/utility.h"
+#include "library/common/bridge/utility.h"
 
 namespace Envoy {
 namespace Platform {
@@ -10,16 +10,16 @@ namespace {
 envoy_data c_kv_store_read(envoy_data key, const void* context) {
   auto kv_store = *static_cast<KeyValueStoreSharedPtr*>(const_cast<void*>(context));
 
-  auto value = kv_store->read(Data::Utility::copyToString(key));
+  auto value = kv_store->read(Bridge::Utility::copyToString(key));
   release_envoy_data(key);
 
-  return value.has_value() ? Data::Utility::copyToBridgeData(value.value()) : envoy_nodata;
+  return value.has_value() ? Bridge::Utility::copyToBridgeData(value.value()) : envoy_nodata;
 }
 
 void c_kv_store_save(envoy_data key, envoy_data value, const void* context) {
   auto kv_store = *static_cast<KeyValueStoreSharedPtr*>(const_cast<void*>(context));
 
-  kv_store->save(Data::Utility::copyToString(key), Data::Utility::copyToString(value));
+  kv_store->save(Bridge::Utility::copyToString(key), Bridge::Utility::copyToString(value));
   release_envoy_data(key);
   release_envoy_data(value);
 }
@@ -27,7 +27,7 @@ void c_kv_store_save(envoy_data key, envoy_data value, const void* context) {
 void c_kv_store_remove(envoy_data key, const void* context) {
   auto kv_store = *static_cast<KeyValueStoreSharedPtr*>(const_cast<void*>(context));
 
-  kv_store->remove(Data::Utility::copyToString(key));
+  kv_store->remove(Bridge::Utility::copyToString(key));
   release_envoy_data(key);
 }
