+++ mobile/library/common/bridge/BUILD
@@ -12,9 +12,12 @@ envoy_cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//library/common:engine_types_lib",
-        "//library/common/data:utility_lib",
+        "//library/common/buffer:bridge_fragment_lib",
         "//library/common/types:c_types_lib",
+        "@envoy//envoy/buffer:buffer_interface",
         "@envoy//envoy/http:codes_interface",
+        "@envoy//source/common/buffer:buffer_lib",
+        "@envoy//source/common/common:empty_string",
         "@envoy//source/common/http:codes_lib",
     ],
 )
