--- mobile/library/common/data/BUILD
@@ -1,19 +1,0 @@
-load("@envoy//bazel:envoy_build_system.bzl", "envoy_cc_library", "envoy_mobile_package")
-
-licenses(["notice"])  # Apache 2
-
-envoy_mobile_package()
-
-envoy_cc_library(
-    name = "utility_lib",
-    srcs = ["utility.cc"],
-    hdrs = ["utility.h"],
-    repository = "@envoy",
-    deps = [
-        "//library/common/buffer:bridge_fragment_lib",
-        "//library/common/types:c_types_lib",
-        "@envoy//envoy/buffer:buffer_interface",
-        "@envoy//source/common/buffer:buffer_lib",
-        "@envoy//source/common/common:empty_string",
-    ],
-)
