+++ mobile/library/common/extensions/filters/http/network_configuration/filter.cc
@@ -7,7 +7,7 @@
 #include "source/common/network/filter_state_proxy_info.h"
 
 #include "library/common/api/external.h"
-#include "library/common/data/utility.h"
+#include "library/common/bridge//utility.h"
 #include "library/common/http/header_utility.h"
 #include "library/common/types/c_types.h"
