+++ mobile/library/common/extensions/filters/http/platform_bridge/BUILD
@@ -29,7 +29,6 @@ envoy_cc_extension(
         ":filter_cc_proto",
         "//library/common/api:external_api_lib",
         "//library/common/bridge:utility_lib",
-        "//library/common/data:utility_lib",
         "//library/common/http:header_utility_lib",
         "//library/common/stream_info:extra_stream_info_lib",
         "//library/common/types:c_types_lib",
