+++ mobile/library/common/http/BUILD
@@ -14,7 +14,6 @@ envoy_cc_library(
         "//library/common:engine_types_lib",
         "//library/common/bridge:utility_lib",
         "//library/common/buffer:bridge_fragment_lib",
-        "//library/common/data:utility_lib",
         "//library/common/event:provisional_dispatcher_lib",
         "//library/common/extensions/filters/http/local_error:local_error_filter_lib",
         "//library/common/extensions/filters/http/network_configuration:network_configuration_filter_lib",
@@ -53,7 +52,7 @@ envoy_cc_library(
     hdrs = ["header_utility.h"],
     repository = "@envoy",
     deps = [
-        "//library/common/data:utility_lib",
+        "//library/common/bridge:utility_lib",
         "//library/common/types:c_types_lib",
         "@envoy//envoy/buffer:buffer_interface",
         "@envoy//envoy/http:filter_interface",
