+++ mobile/library/common/logger/BUILD
@@ -13,7 +13,6 @@ envoy_cc_library(
         "//library/common:engine_types_lib",
         "//library/common/api:external_api_lib",
         "//library/common/bridge:utility_lib",
-        "//library/common/data:utility_lib",
         "//library/common/types:c_types_lib",
         "@envoy//source/common/common:macros",
         "@envoy//source/common/common:minimal_logger_lib",
