+++ mobile/library/common/stats/BUILD
@@ -12,7 +12,7 @@ envoy_cc_library(
     hdrs = ["utility.h"],
     repository = "@envoy",
     deps = [
-        "//library/common/data:utility_lib",
+        "//library/common/bridge:utility_lib",
         "@envoy//source/common/stats:utility_lib",
     ],
 )
