+++ mobile/library/common/stats/utility.cc
@@ -1,7 +1,7 @@
 // From upstream Envoy
 #include "source/common/stats/utility.h"
 
-#include "library/common/data/utility.h"
+#include "library/common/bridge//utility.h"
 #include "library/common/types/c_types.h"
 
 namespace Envoy {
@@ -12,8 +12,8 @@ Stats::StatNameTagVector transformToStatNameTagVector(envoy_stats_tags tags,
                                                       Stats::StatNameSetPtr& stat_name_set) {
   Stats::StatNameTagVector transformed_tags;
   for (envoy_map_size_t i = 0; i < tags.length; i++) {
-    std::string key(Data::Utility::copyToString(tags.entries[i].key));
-    std::string val(Data::Utility::copyToString(tags.entries[i].value));
+    std::string key(Bridge::Utility::copyToString(tags.entries[i].key));
+    std::string val(Bridge::Utility::copyToString(tags.entries[i].value));
 
     transformed_tags.push_back({stat_name_set->add(key), stat_name_set->add(val)});
   }
