+++ mobile/library/jni/BUILD
@@ -81,7 +81,6 @@ envoy_cc_library(
         "//library/common:internal_engine_lib",
         "//library/common/api:c_types",
         "//library/common/bridge:utility_lib",
-        "//library/common/data:utility_lib",
         "//library/common/extensions/cert_validator/platform_bridge:c_types_lib",
         "//library/common/extensions/key_value/platform:config",
         "//library/common/types:managed_types_lib",
@@ -105,7 +104,7 @@ envoy_cc_library(
     repository = "@envoy",
     deps = [
         ":jni_utility_lib",
-        "//library/common/data:utility_lib",
+        "//library/common/bridge:utility_lib",
         "//library/common/extensions/cert_validator/platform_bridge:c_types_lib",
         "//library/common/types:c_types_lib",
         "//library/jni/import:jni_import_lib",
@@ -205,7 +204,7 @@ envoy_cc_library(
         "@envoy//bazel:android": [
             ":jni_support_lib",
             ":jni_utility_lib",
-            "//library/common/data:utility_lib",
+            "//library/common/bridge:utility_lib",
             "//library/jni/import:jni_import_lib",
         ],
         "//conditions:default": [],
