+++ mobile/library/jni/android_jni_utility.cc
@@ -3,7 +3,7 @@
 #include "source/common/common/assert.h"
 
 #if defined(__ANDROID_API__)
-#include "library/common/data/utility.h"
+#include "library/common/bridge/utility.h"
 #include "library/jni/import/jni_import.h"
 #include "library/jni/jni_support.h"
 #include "library/jni/jni_utility.h"
@@ -14,7 +14,7 @@ namespace JNI {
 
 bool isCleartextPermitted(absl::string_view hostname) {
 #if defined(__ANDROID_API__)
-  envoy_data host = Envoy::Data::Utility::copyToBridgeData(hostname);
+  envoy_data host = Envoy::Bridge::Utility::copyToBridgeData(hostname);
   JniHelper jni_helper(getEnv());
   LocalRefUniquePtr<jstring> java_host = envoyDataToJavaString(jni_helper, host);
   LocalRefUniquePtr<jclass> jcls_AndroidNetworkLibrary =
