+++ mobile/library/objective-c/BUILD
@@ -60,7 +60,6 @@ envoy_objc_library(
         "//library/common:internal_engine_lib",
         "//library/common/api:c_types",
         "//library/common/bridge:utility_lib",
-        "//library/common/data:utility_lib",
         "//library/common/http:header_utility_lib",
         "//library/common/network:apple_platform_cert_verifier",
         "//library/common/network:apple_proxy_resolution_lib",
