+++ mobile/test/cc/unit/envoy_config_test.cc
@@ -15,7 +15,7 @@
 #include "gtest/gtest.h"
 #include "library/cc/engine_builder.h"
 #include "library/common/api/external.h"
-#include "library/common/data/utility.h"
+#include "library/common/bridge//utility.h"
 
 #if defined(__APPLE__)
 #include "source/extensions/network/dns_resolver/apple/apple_dns_impl.h"
@@ -521,7 +521,7 @@ TEST(TestConfig, DISABLED_StringAccessors) {
   EXPECT_EQ(0, accessor->count());
   envoy_data data = c_accessor->get_string(c_accessor->context);
   EXPECT_EQ(1, accessor->count());
-  EXPECT_EQ(data_string, Data::Utility::copyToString(data));
+  EXPECT_EQ(data_string, Bridge::Utility::copyToString(data));
   release_envoy_data(data);
 }
