+++ mobile/test/common/bridge/BUILD
@@ -10,5 +10,7 @@ envoy_cc_test(
     repository = "@envoy",
     deps = [
         "//library/common/bridge:utility_lib",
+        "//library/common/types:c_types_lib",
+        "@envoy//source/common/buffer:buffer_lib",
     ],
 )
