+++ mobile/test/common/extensions/filters/http/network_configuration/BUILD
@@ -15,7 +15,7 @@ envoy_extension_cc_test(
     repository = "@envoy",
     deps = [
         "//library/common/api:external_api_lib",
-        "//library/common/data:utility_lib",
+        "//library/common/bridge:utility_lib",
         "//library/common/extensions/filters/http/network_configuration:config",
         "//library/common/extensions/filters/http/network_configuration:filter_cc_proto",
         "//library/common/network:proxy_api_lib",
