+++ mobile/test/common/extensions/filters/http/platform_bridge/BUILD
@@ -15,7 +15,7 @@ envoy_extension_cc_test(
     repository = "@envoy",
     deps = [
         "//library/common/api:external_api_lib",
-        "//library/common/data:utility_lib",
+        "//library/common/bridge:utility_lib",
         "//library/common/extensions/filters/http/platform_bridge:config",
         "//library/common/extensions/filters/http/platform_bridge:filter_cc_proto",
         "@envoy//test/mocks/event:event_mocks",
