+++ mobile/test/common/extensions/filters/http/platform_bridge/platform_bridge_filter_test.cc
@@ -5,7 +5,7 @@
 
 #include "gtest/gtest.h"
 #include "library/common/api/external.h"
-#include "library/common/data/utility.h"
+#include "library/common/bridge/utility.h"
 #include "library/common/extensions/filters/http/platform_bridge/filter.h"
 #include "library/common/extensions/filters/http/platform_bridge/filter.pb.h"
 
@@ -28,8 +28,8 @@ envoy_headers make_envoy_headers(std::vector<std::pair<std::string, std::string>
   new_headers.entries = headers;
 
   for (const auto& pair : pairs) {
-    envoy_data key = Data::Utility::copyToBridgeData(pair.first);
-    envoy_data value = Data::Utility::copyToBridgeData(pair.second);
+    envoy_data key = Bridge::Utility::copyToBridgeData(pair.first);
+    envoy_data value = Bridge::Utility::copyToBridgeData(pair.second);
 
     new_headers.entries[new_headers.length] = {key, value};
     new_headers.length++;
@@ -198,8 +198,8 @@ TEST_F(PlatformBridgeFilterTest, BasicContinueOnRequestHeaders) {
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_TRUE(end_stream);
     invocations->on_request_headers_calls++;
     return {kEnvoyFilterHeadersStatusContinue, c_headers};
@@ -230,8 +230,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnData) {
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
@@ -240,7 +240,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnData) {
   platform_filter.on_request_data = [](envoy_data c_data, bool end_stream, envoy_stream_intel,
                                        const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
-    EXPECT_EQ(Data::Utility::copyToString(c_data), "request body");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data), "request body");
     EXPECT_TRUE(end_stream);
     invocations->on_request_data_calls++;
     envoy_headers* modified_headers =
@@ -285,8 +285,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnResumeDecoding)
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
@@ -297,8 +297,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnResumeDecoding)
          bool end_stream, envoy_stream_intel, const void* context) -> envoy_filter_resume_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(pending_headers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].value), "test.code");
     EXPECT_EQ(pending_data, nullptr);
     EXPECT_EQ(pending_trailers, nullptr);
     EXPECT_FALSE(end_stream);
@@ -351,8 +351,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnResumeDecodingW
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
@@ -363,8 +363,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnResumeDecodingW
          bool end_stream, envoy_stream_intel, const void* context) -> envoy_filter_resume_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(pending_headers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].value), "test.code");
     EXPECT_EQ(pending_data, nullptr);
     EXPECT_EQ(pending_trailers, nullptr);
     EXPECT_FALSE(end_stream);
@@ -377,7 +377,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnResumeDecodingW
 
     Buffer::OwnedImpl final_buffer = Buffer::OwnedImpl("C");
     envoy_data* modified_data = static_cast<envoy_data*>(safe_malloc(sizeof(envoy_data)));
-    *modified_data = Data::Utility::toBridgeData(final_buffer);
+    *modified_data = Bridge::Utility::toBridgeData(final_buffer);
     return {kEnvoyFilterResumeStatusResumeIteration, modified_headers, modified_data, nullptr};
   };
 
@@ -421,8 +421,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnResumeDecodingW
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
@@ -433,8 +433,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnResumeDecodingW
          bool end_stream, envoy_stream_intel, const void* context) -> envoy_filter_resume_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(pending_headers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].value), "test.code");
     EXPECT_EQ(pending_data, nullptr);
     EXPECT_EQ(pending_trailers, nullptr);
     EXPECT_FALSE(end_stream);
@@ -498,8 +498,8 @@ TEST_F(PlatformBridgeFilterTest, AsyncResumeDecodingIsNoopAfterPreviousResume) {
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
@@ -508,7 +508,7 @@ TEST_F(PlatformBridgeFilterTest, AsyncResumeDecodingIsNoopAfterPreviousResume) {
   platform_filter.on_request_data = [](envoy_data c_data, bool end_stream, envoy_stream_intel,
                                        const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
-    EXPECT_EQ(Data::Utility::copyToString(c_data), "request body");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data), "request body");
     EXPECT_TRUE(end_stream);
     invocations->on_request_data_calls++;
     envoy_headers* modified_headers =
@@ -566,7 +566,7 @@ TEST_F(PlatformBridgeFilterTest, BasicContinueOnRequestData) {
   platform_filter.on_request_data = [](envoy_data c_data, bool end_stream, envoy_stream_intel,
                                        const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
-    EXPECT_EQ(Data::Utility::copyToString(c_data), "request body");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data), "request body");
     EXPECT_TRUE(end_stream);
     invocations->on_request_data_calls++;
     return {kEnvoyFilterDataStatusContinue, c_data, nullptr};
@@ -596,7 +596,7 @@ TEST_F(PlatformBridgeFilterTest, StopAndBufferOnRequestData) {
                                        const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     std::string expected_data[3] = {"A", "AB", "ABC"};
-    EXPECT_EQ(Data::Utility::copyToString(c_data),
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data),
               expected_data[invocations->on_request_data_calls++]);
     EXPECT_FALSE(end_stream);
     release_envoy_data(c_data);
@@ -667,7 +667,7 @@ TEST_F(PlatformBridgeFilterTest, BasicError) {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     invocations->on_error_calls++;
     EXPECT_EQ(c_error.error_code, ENVOY_UNDEFINED_ERROR);
-    EXPECT_EQ(Data::Utility::copyToString(c_error.message), "busted");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_error.message), "busted");
     EXPECT_EQ(c_error.attempt_count, 1);
     release_envoy_error(c_error);
   };
@@ -707,17 +707,17 @@ TEST_F(PlatformBridgeFilterTest, StopAndBufferThenResumeOnRequestData) {
     envoy_filter_data_status return_status;
 
     if (invocations->on_request_data_calls == 0) {
-      EXPECT_EQ(Data::Utility::copyToString(c_data), "A");
+      EXPECT_EQ(Bridge::Utility::copyToString(c_data), "A");
       EXPECT_FALSE(end_stream);
 
       return_status.status = kEnvoyFilterDataStatusStopIterationAndBuffer;
       return_status.data = envoy_nodata;
       return_status.pending_headers = nullptr;
     } else {
-      EXPECT_EQ(Data::Utility::copyToString(c_data), "AB");
+      EXPECT_EQ(Bridge::Utility::copyToString(c_data), "AB");
       EXPECT_FALSE(end_stream);
       Buffer::OwnedImpl final_buffer = Buffer::OwnedImpl("C");
-      envoy_data final_data = Data::Utility::toBridgeData(final_buffer);
+      envoy_data final_data = Bridge::Utility::toBridgeData(final_buffer);
 
       return_status.status = kEnvoyFilterDataStatusResumeIteration;
       return_status.data = final_data;
@@ -773,8 +773,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnData)
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
@@ -786,17 +786,17 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnData)
     envoy_filter_data_status return_status;
 
     if (invocations->on_request_data_calls == 0) {
-      EXPECT_EQ(Data::Utility::copyToString(c_data), "A");
+      EXPECT_EQ(Bridge::Utility::copyToString(c_data), "A");
       EXPECT_FALSE(end_stream);
 
       return_status.status = kEnvoyFilterDataStatusStopIterationAndBuffer;
       return_status.data = envoy_nodata;
       return_status.pending_headers = nullptr;
     } else {
-      EXPECT_EQ(Data::Utility::copyToString(c_data), "AB");
+      EXPECT_EQ(Bridge::Utility::copyToString(c_data), "AB");
       EXPECT_TRUE(end_stream);
       Buffer::OwnedImpl final_buffer = Buffer::OwnedImpl("C");
-      envoy_data final_data = Data::Utility::toBridgeData(final_buffer);
+      envoy_data final_data = Bridge::Utility::toBridgeData(final_buffer);
       envoy_headers* modified_headers =
           static_cast<envoy_headers*>(safe_malloc(sizeof(envoy_headers)));
       *modified_headers =
@@ -867,7 +867,7 @@ TEST_F(PlatformBridgeFilterTest, StopNoBufferOnRequestData) {
                                        const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     std::string expected_data[3] = {"A", "B", "C"};
-    EXPECT_EQ(Data::Utility::copyToString(c_data),
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data),
               expected_data[invocations->on_request_data_calls++]);
     EXPECT_FALSE(end_stream);
     release_envoy_data(c_data);
@@ -906,8 +906,8 @@ TEST_F(PlatformBridgeFilterTest, BasicContinueOnRequestTrailers) {
                                            const void* context) -> envoy_filter_trailers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_trailers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
     invocations->on_request_trailers_calls++;
     return {kEnvoyFilterTrailersStatusContinue, c_trailers, nullptr, nullptr};
   };
@@ -937,8 +937,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnTrail
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
@@ -948,7 +948,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnTrail
                                        const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     std::string expected_data[2] = {"A", "AB"};
-    EXPECT_EQ(Data::Utility::copyToString(c_data),
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data),
               expected_data[invocations->on_request_data_calls]);
     EXPECT_FALSE(end_stream);
     release_envoy_data(c_data);
@@ -959,12 +959,12 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnTrail
                                            const void* context) -> envoy_filter_trailers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_trailers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
 
     Buffer::OwnedImpl final_buffer = Buffer::OwnedImpl("C");
     envoy_data* modified_data = static_cast<envoy_data*>(safe_malloc(sizeof(envoy_data)));
-    *modified_data = Data::Utility::toBridgeData(final_buffer);
+    *modified_data = Bridge::Utility::toBridgeData(final_buffer);
     envoy_headers* modified_headers =
         static_cast<envoy_headers*>(safe_malloc(sizeof(envoy_headers)));
     *modified_headers = make_envoy_headers({{":authority", "test.code"}, {"content-length", "1"}});
@@ -1037,8 +1037,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnResum
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
@@ -1048,7 +1048,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnResum
                                        const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     std::string expected_data[2] = {"A", "AB"};
-    EXPECT_EQ(Data::Utility::copyToString(c_data),
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data),
               expected_data[invocations->on_request_data_calls]);
     EXPECT_FALSE(end_stream);
     release_envoy_data(c_data);
@@ -1059,8 +1059,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnResum
                                            const void* context) -> envoy_filter_trailers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_trailers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
     release_envoy_headers(c_trailers);
     invocations->on_request_trailers_calls++;
     return {kEnvoyFilterTrailersStatusStopIteration, envoy_noheaders, nullptr, nullptr};
@@ -1070,12 +1070,12 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnResum
          bool end_stream, envoy_stream_intel, const void* context) -> envoy_filter_resume_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(pending_headers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].value), "test.code");
-    EXPECT_EQ(Data::Utility::copyToString(*pending_data), "AB");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(*pending_data), "AB");
     EXPECT_EQ(pending_trailers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_trailers->entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(pending_trailers->entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_trailers->entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_trailers->entries[0].value), "test trailer");
     EXPECT_TRUE(end_stream);
 
     envoy_headers* modified_headers =
@@ -1085,7 +1085,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenResumeOnResum
     release_envoy_headers(*pending_headers);
     Buffer::OwnedImpl final_buffer = Buffer::OwnedImpl("C");
     envoy_data* modified_data = static_cast<envoy_data*>(safe_malloc(sizeof(envoy_data)));
-    *modified_data = Data::Utility::toBridgeData(final_buffer);
+    *modified_data = Bridge::Utility::toBridgeData(final_buffer);
     release_envoy_data(*pending_data);
     envoy_headers* modified_trailers =
         static_cast<envoy_headers*>(safe_malloc(sizeof(envoy_headers)));
@@ -1177,8 +1177,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenDontResumeOnR
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
@@ -1188,7 +1188,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenDontResumeOnR
                                        const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     std::string expected_data[2] = {"A", "AB"};
-    EXPECT_EQ(Data::Utility::copyToString(c_data),
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data),
               expected_data[invocations->on_request_data_calls]);
     EXPECT_FALSE(end_stream);
     release_envoy_data(c_data);
@@ -1199,8 +1199,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenDontResumeOnR
                                            const void* context) -> envoy_filter_trailers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_trailers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
     release_envoy_headers(c_trailers);
     invocations->on_request_trailers_calls++;
     return {kEnvoyFilterTrailersStatusStopIteration, envoy_noheaders, nullptr, nullptr};
@@ -1210,12 +1210,12 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenBufferThenDontResumeOnR
          bool end_stream, envoy_stream_intel, const void* context) -> envoy_filter_resume_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(pending_headers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].value), "test.code");
-    EXPECT_EQ(Data::Utility::copyToString(*pending_data), "AB");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(*pending_data), "AB");
     EXPECT_EQ(pending_trailers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_trailers->entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(pending_trailers->entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_trailers->entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_trailers->entries[0].value), "test trailer");
     EXPECT_TRUE(end_stream);
 
     release_envoy_headers(*pending_headers);
@@ -1291,8 +1291,8 @@ TEST_F(PlatformBridgeFilterTest, BasicContinueOnResponseHeaders) {
                                            const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_TRUE(end_stream);
     invocations->on_response_headers_calls++;
     return {kEnvoyFilterHeadersStatusContinue, c_headers};
@@ -1323,8 +1323,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenResumeOnData) {
                                            const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_response_headers_calls++;
     release_envoy_headers(c_headers);
@@ -1333,7 +1333,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenResumeOnData) {
   platform_filter.on_response_data = [](envoy_data c_data, bool end_stream, envoy_stream_intel,
                                         const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
-    EXPECT_EQ(Data::Utility::copyToString(c_data), "response body");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data), "response body");
     EXPECT_TRUE(end_stream);
     invocations->on_response_data_calls++;
     envoy_headers* modified_headers =
@@ -1378,8 +1378,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenResumeOnResumeEncoding
                                            const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_response_headers_calls++;
     release_envoy_headers(c_headers);
@@ -1390,8 +1390,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenResumeOnResumeEncoding
          bool end_stream, envoy_stream_intel, const void* context) -> envoy_filter_resume_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(pending_headers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].value), "test.code");
     EXPECT_EQ(pending_data, nullptr);
     EXPECT_EQ(pending_trailers, nullptr);
     EXPECT_FALSE(end_stream);
@@ -1444,8 +1444,8 @@ TEST_F(PlatformBridgeFilterTest, AsyncResumeEncodingIsNoopAfterPreviousResume) {
                                            const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_response_headers_calls++;
     release_envoy_headers(c_headers);
@@ -1454,7 +1454,7 @@ TEST_F(PlatformBridgeFilterTest, AsyncResumeEncodingIsNoopAfterPreviousResume) {
   platform_filter.on_response_data = [](envoy_data c_data, bool end_stream, envoy_stream_intel,
                                         const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
-    EXPECT_EQ(Data::Utility::copyToString(c_data), "response body");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data), "response body");
     EXPECT_TRUE(end_stream);
     invocations->on_response_data_calls++;
     envoy_headers* modified_headers =
@@ -1514,8 +1514,8 @@ TEST_F(PlatformBridgeFilterTest, AsyncResumeEncodingIsNoopAfterFilterIsPendingDe
                                            const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_response_headers_calls++;
     release_envoy_headers(c_headers);
@@ -1576,7 +1576,7 @@ TEST_F(PlatformBridgeFilterTest, BasicContinueOnResponseData) {
   platform_filter.on_response_data = [](envoy_data c_data, bool end_stream, envoy_stream_intel,
                                         const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
-    EXPECT_EQ(Data::Utility::copyToString(c_data), "response body");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data), "response body");
     EXPECT_TRUE(end_stream);
     invocations->on_response_data_calls++;
     return {kEnvoyFilterDataStatusContinue, c_data, nullptr};
@@ -1606,7 +1606,7 @@ TEST_F(PlatformBridgeFilterTest, StopAndBufferOnResponseData) {
                                         const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     std::string expected_data[3] = {"A", "AB", "ABC"};
-    EXPECT_EQ(Data::Utility::copyToString(c_data),
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data),
               expected_data[invocations->on_response_data_calls++]);
     EXPECT_FALSE(end_stream);
     release_envoy_data(c_data);
@@ -1662,17 +1662,17 @@ TEST_F(PlatformBridgeFilterTest, StopAndBufferThenResumeOnResponseData) {
     envoy_filter_data_status return_status;
 
     if (invocations->on_response_data_calls == 0) {
-      EXPECT_EQ(Data::Utility::copyToString(c_data), "A");
+      EXPECT_EQ(Bridge::Utility::copyToString(c_data), "A");
       EXPECT_FALSE(end_stream);
 
       return_status.status = kEnvoyFilterDataStatusStopIterationAndBuffer;
       return_status.data = envoy_nodata;
       return_status.pending_headers = nullptr;
     } else {
-      EXPECT_EQ(Data::Utility::copyToString(c_data), "AB");
+      EXPECT_EQ(Bridge::Utility::copyToString(c_data), "AB");
       EXPECT_FALSE(end_stream);
       Buffer::OwnedImpl final_buffer = Buffer::OwnedImpl("C");
-      envoy_data final_data = Data::Utility::toBridgeData(final_buffer);
+      envoy_data final_data = Bridge::Utility::toBridgeData(final_buffer);
 
       return_status.status = kEnvoyFilterDataStatusResumeIteration;
       return_status.data = final_data;
@@ -1728,8 +1728,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnData
                                            const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_response_headers_calls++;
     release_envoy_headers(c_headers);
@@ -1741,17 +1741,17 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnData
     envoy_filter_data_status return_status;
 
     if (invocations->on_response_data_calls == 0) {
-      EXPECT_EQ(Data::Utility::copyToString(c_data), "A");
+      EXPECT_EQ(Bridge::Utility::copyToString(c_data), "A");
       EXPECT_FALSE(end_stream);
 
       return_status.status = kEnvoyFilterDataStatusStopIterationAndBuffer;
       return_status.data = envoy_nodata;
       return_status.pending_headers = nullptr;
     } else {
-      EXPECT_EQ(Data::Utility::copyToString(c_data), "AB");
+      EXPECT_EQ(Bridge::Utility::copyToString(c_data), "AB");
       EXPECT_TRUE(end_stream);
       Buffer::OwnedImpl final_buffer = Buffer::OwnedImpl("C");
-      envoy_data final_data = Data::Utility::toBridgeData(final_buffer);
+      envoy_data final_data = Bridge::Utility::toBridgeData(final_buffer);
       envoy_headers* modified_headers =
           static_cast<envoy_headers*>(safe_malloc(sizeof(envoy_headers)));
       *modified_headers = make_envoy_headers({{":status", "test.code"}, {"content-length", "1"}});
@@ -1821,7 +1821,7 @@ TEST_F(PlatformBridgeFilterTest, StopNoBufferOnResponseData) {
                                         const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     std::string expected_data[3] = {"A", "B", "C"};
-    EXPECT_EQ(Data::Utility::copyToString(c_data),
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data),
               expected_data[invocations->on_response_data_calls++]);
     EXPECT_FALSE(end_stream);
     release_envoy_data(c_data);
@@ -1860,8 +1860,8 @@ TEST_F(PlatformBridgeFilterTest, BasicContinueOnResponseTrailers) {
                                             const void* context) -> envoy_filter_trailers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_trailers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
     invocations->on_response_trailers_calls++;
     return {kEnvoyFilterTrailersStatusContinue, c_trailers, nullptr, nullptr};
   };
@@ -1891,8 +1891,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnTrai
                                            const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_response_headers_calls++;
     release_envoy_headers(c_headers);
@@ -1902,7 +1902,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnTrai
                                         const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     std::string expected_data[2] = {"A", "AB"};
-    EXPECT_EQ(Data::Utility::copyToString(c_data),
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data),
               expected_data[invocations->on_response_data_calls]);
     EXPECT_FALSE(end_stream);
     release_envoy_data(c_data);
@@ -1913,12 +1913,12 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnTrai
                                             const void* context) -> envoy_filter_trailers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_trailers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
 
     Buffer::OwnedImpl final_buffer = Buffer::OwnedImpl("C");
     envoy_data* modified_data = static_cast<envoy_data*>(safe_malloc(sizeof(envoy_data)));
-    *modified_data = Data::Utility::toBridgeData(final_buffer);
+    *modified_data = Bridge::Utility::toBridgeData(final_buffer);
     envoy_headers* modified_headers =
         static_cast<envoy_headers*>(safe_malloc(sizeof(envoy_headers)));
     *modified_headers = make_envoy_headers({{":status", "test.code"}, {"content-length", "1"}});
@@ -1991,8 +1991,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnResu
                                            const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_response_headers_calls++;
     release_envoy_headers(c_headers);
@@ -2002,7 +2002,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnResu
                                         const void* context) -> envoy_filter_data_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     std::string expected_data[2] = {"A", "AB"};
-    EXPECT_EQ(Data::Utility::copyToString(c_data),
+    EXPECT_EQ(Bridge::Utility::copyToString(c_data),
               expected_data[invocations->on_response_data_calls]);
     EXPECT_FALSE(end_stream);
     release_envoy_data(c_data);
@@ -2013,8 +2013,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnResu
                                             const void* context) -> envoy_filter_trailers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_trailers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_trailers.entries[0].value), "test trailer");
     release_envoy_headers(c_trailers);
     invocations->on_response_trailers_calls++;
     return {kEnvoyFilterTrailersStatusStopIteration, envoy_noheaders, nullptr, nullptr};
@@ -2024,12 +2024,12 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnResu
          bool end_stream, envoy_stream_intel, const void* context) -> envoy_filter_resume_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(pending_headers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].key), ":status");
-    EXPECT_EQ(Data::Utility::copyToString(pending_headers->entries[0].value), "test.code");
-    EXPECT_EQ(Data::Utility::copyToString(*pending_data), "AB");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].key), ":status");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_headers->entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(*pending_data), "AB");
     EXPECT_EQ(pending_trailers->length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(pending_trailers->entries[0].key), "x-test-trailer");
-    EXPECT_EQ(Data::Utility::copyToString(pending_trailers->entries[0].value), "test trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_trailers->entries[0].key), "x-test-trailer");
+    EXPECT_EQ(Bridge::Utility::copyToString(pending_trailers->entries[0].value), "test trailer");
     EXPECT_TRUE(end_stream);
 
     envoy_headers* modified_headers =
@@ -2039,7 +2039,7 @@ TEST_F(PlatformBridgeFilterTest, StopOnResponseHeadersThenBufferThenResumeOnResu
     release_envoy_headers(*pending_headers);
     Buffer::OwnedImpl final_buffer = Buffer::OwnedImpl("C");
     envoy_data* modified_data = static_cast<envoy_data*>(safe_malloc(sizeof(envoy_data)));
-    *modified_data = Data::Utility::toBridgeData(final_buffer);
+    *modified_data = Bridge::Utility::toBridgeData(final_buffer);
     release_envoy_data(*pending_data);
     envoy_headers* modified_trailers =
         static_cast<envoy_headers*>(safe_malloc(sizeof(envoy_headers)));
@@ -2131,8 +2131,8 @@ TEST_F(PlatformBridgeFilterTest, StopOnRequestHeadersThenResumeOnResumeDecodingP
                                           const void* context) -> envoy_filter_headers_status {
     filter_invocations* invocations = static_cast<filter_invocations*>(const_cast<void*>(context));
     EXPECT_EQ(c_headers.length, 1);
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].key), ":authority");
-    EXPECT_EQ(Data::Utility::copyToString(c_headers.entries[0].value), "test.code");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].key), ":authority");
+    EXPECT_EQ(Bridge::Utility::copyToString(c_headers.entries[0].value), "test.code");
     EXPECT_FALSE(end_stream);
     invocations->on_request_headers_calls++;
     release_envoy_headers(c_headers);
