+++ mobile/test/common/http/BUILD
@@ -31,7 +31,6 @@ envoy_cc_test(
     repository = "@envoy",
     deps = [
         "//library/common/bridge:utility_lib",
-        "//library/common/data:utility_lib",
         "//library/common/http:header_utility_lib",
         "//library/common/types:c_types_lib",
         "@envoy//source/common/buffer:buffer_lib",
