+++ mobile/test/common/http/filters/test_event_tracker/BUILD
@@ -28,7 +28,6 @@ envoy_cc_library(
         "//library/common/api:c_types",
         "//library/common/api:external_api_lib",
         "//library/common/bridge:utility_lib",
-        "//library/common/data:utility_lib",
         "@envoy//source/common/common:assert_lib",
         "@envoy//source/extensions/filters/http/common:pass_through_filter_lib",
     ],
