+++ mobile/test/common/integration/client_integration_test.cc
@@ -16,7 +16,7 @@
 #include "test/test_common/test_random_generator.h"
 
 #include "extension_registry.h"
-#include "library/common/data/utility.h"
+#include "library/common/bridge/utility.h"
 #include "library/common/http/header_utility.h"
 #include "library/common/internal_engine.h"
 #include "library/common/network/proxy_settings.h"
