+++ mobile/test/common/logger/logger_delegate_test.cc
@@ -1,7 +1,7 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include "library/common/api/external.h"
-#include "library/common/data/utility.h"
+#include "library/common/bridge/utility.h"
 #include "library/common/logger/logger_delegate.h"
 
 using testing::_;
